Added TOML stuffs
Added a couple of scripts to the quest manager and created a few to read from .toml files. These/this format will probably be used for saving and loading any save data for the game. Quests have tasks which are loaded from one of these .toml files and then put into a list of temporary tasks. These tasks will then be put into another type of object later, in order to use functions from them to interact with the quest manager itself.