Fixed village seasonal effect issue, added cylinder
Fixed the issue where placing an object in the village would be detected as placing it in one of the seasons. The fix is a little ugly (each season has a copy of the village ground, and raycasts it first),will improve after testing.

Added a cylinder to the summer season for testing.