Uniform code style, integration of QuestEventManager
Formatted all of our scripts (hopefully) to use the same uniform code style.

Moved some of the events used internally by some of our scripts to an InternalEvents file and namespace. Hopefully this will make our event system a little bit easier to understand and use.

Worked on integrating QuestEventManager into the rest of the game, mostly by adding event handlers to it.

Refactored the code that generates the trigger and OnComplete predicates. Additionally, moved the events that were defined in QuestEventManager to Events. Also added Store and Retrieve InteractionKinds.

Removed SummerLeaves script, it seems to be unused.