Cleanup, fixed inventory system, fixed laser pointer issue
Inventory system now properly fires all the events it should, previously the Object[Accepted/Rejected]ByInventoryTargetEvent(s) were never used. This led to the inventory manager's state not updating when objects were added/removed manually.

Also moved AllScenesLoadedEvent to InternalEvents.

Cleaned up and improved PlayerLocationTracker.

Fixed an issue in the SetVisibility method of IUILaserPointer where we never checked if pointer and hitPoint were not null before attempting to call methods on them (they will be null if the controller is not connected)