Objects can now be picked up from farther away issue #21
Before the controller had to be in direct contact with the objects in order to pick them up, now the user can be quite the distance away from the objects and still be able to pick them up. The z-axis was increased in size on both controller's colliders. In order to still allow teleporting to be functional, the center of the collider on the y-axis also had to be moved. It has not been tested whether the current values are appropriate.
Original:
Center: 0, -0.04, 0.02
Size: 0.14, 0.07, 0.05
Current:
Center: 0, -0.5, 0.02
Size: 0.14, 0.07, 1